+++ example/lib/utils/message.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+void showMessage(BuildContext context, String message, {Color? backgroundColor}) {
+  ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+    content: Text(message),
+    backgroundColor: backgroundColor,
+  ));
+}