+++ example/lib/widgets/pages/submit_form_builder.page.dart
@@ -1,3 +1,4 @@
+import 'package:example/utils/message.dart';
 import 'package:fetcher/fetcher.dart';
 import 'package:flutter/material.dart';
 
@@ -14,9 +15,9 @@ class _SubmitFormBuilderPageState extends State<SubmitFormBuilderPage> {
   @override
   Widget build(BuildContext context) {
     return SubmitFormBuilder<String>(
-      onSuccess: (result) => ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content: Text(result),
-      )),
+      onChanged: () => print('Form changed'),
+      //onUnsavedFormPop: SubmitFormBuilder.ignoreFormPopCallback,    // Uncomment to override behavior set in main.dart
+      onSuccess: (result) => showMessage(context, result),
       builder: (context, validate) {
         return Padding(
           padding: const EdgeInsets.all(15),